requirements.txt
pymongo
pytest
pika
pyredis